requirements.txt
fastapi==0.99.0
boto3
python-dotenv
mangum
openai